refactor: change logic alert toast
