+++ dashboard/pages/object/index.vue
@@ -104,7 +104,7 @@
       text-confirm="Delete"
       type-modal="danger"
       @close="state.isOpenDelete = false"
-      @confirm="deleteData()"
+      @confirm="confirmDelete()"
     />
     <BaseModal
       :open-modal="state.isOpenObject"
@@ -122,6 +122,7 @@
 <script setup lang="ts">
   import { useActivePage } from '@/store/index'
   import { objectHeaders } from '~/common/constant/object'
+  import { deleteData, alertDanger, alertSuccess } from '~/utils'
 
   interface statusColor{
     color: string,
@@ -129,7 +130,6 @@
   }
 
   const router = useRouter()
-  const toast = useToast()
   const state = reactive({
     isOpenDelete: false,
     isOpenObject: false,
@@ -214,41 +214,8 @@
     state.isOpenObject = true
   }
 
-  function deleteData() {
-    let cancelled = false
-    toast.add({
-      icon: 'i-heroicons-exclamation-triangle',
-      title: 'data will be deleted',
-      color: 'yellow',
-      timeout: 2000,
-      callback: async () => {
-        if (!cancelled) {
-          await useDeleteData(urlAPI, state.idItems)
-          state.fetchObject()
-          toast.add({
-            title: 'data successfully deleted',
-            icon: 'i-heroicons-check-circle',
-            timeout: 1000,
-          })
-        }
-      },
-      actions: [
-        {
-          label: 'Undo',
-          variant: 'solid',
-          color: 'orange',
-          click: () => {
-            cancelled = true
-            toast.add({
-              title: 'data cancel deleted',
-              color: 'red',
-              icon: 'i-heroicons-x-circle',
-              timeout: 1000,
-            })
-          },
-        },
-      ],
-    })
+  async function confirmDelete(){
+    deleteData({ urlAPI, id: state.idItems, fetch: state.fetchObject})
     state.isOpenDelete = false
   }
 
@@ -258,12 +225,7 @@
     }
     await useUpdatePatchData(urlAPI, state.idItems, status)
     state.fetchObject()
-    toast.add({
-      icon: 'i-heroicons-exclamation-triangle',
-      title: 'data successfully Update',
-      color: 'green',
-      timeout: 1000,
-    })
+    alertSuccess('data successfully Update')
     state.isOpenObject = false
   }
